feat: Prompt user to install MetaMask and provide download link
This commit adds a prompt to the application when MetaMask is not detected. Instead of just showing an alert, it now displays a confirmation dialog with the option to visit the MetaMask download page. If the user chooses to visit the page, it opens the download link in a new tab.